moves game logic into module, creates additional modules
