+++ src/game_logic/mod.rs
@@ -1,0 +1,5 @@
+pub(crate) mod board;
+mod chain;
+pub mod game;
+mod game_move;
+pub(crate) mod stone;
