+++ src/game_logic/stone.rs
