+++ src/main.rs
@@ -1,5 +1,5 @@
 use anyhow::Result;
-use go::*;
+use go::{game::Game, *};
 
 fn main() -> Result<()> {
     let mut game: Game<RawModeUi> = Game::new_game(9, 9, Default::default());
